git-genpatches: Try to apply patches before generating reverts.
The conflict detection in getclusters() is made by comparing hunk
overlaps and sometimes results in detected conflicts also in cases where
the patch series will apply without problems.

New function apply_clusters() tries to 'git apply' the patches in the
cluster to the temporary branch, and only if that fails starts to
generate revert patches.